deploy_temp/requirements.txt
yfinance>=0.2.8
pandas>=1.5.3
numpy>=1.24.2
ta>=0.10.2
matplotlib>=3.7.1
scikit-learn>=1.2.2
xgboost>=1.7.5
ccxt>=4.0.0
openai>=0.27.8
python-dotenv>=1.0.0
pyyaml>=6.0
MetaTrader5>=5.0.44
pytz>=2022.7.1
joblib>=1.2.0
argparse>=1.4.0
termcolor>=2.3.0
tenacity>=8.2.2
tqdm>=4.65.0
tabulate>=0.9.0 